refactor(EDCO-55)!: Video component refactored
