+++ src/index.ts
@@ -3,7 +3,7 @@ import Feature from './components/Feature/Feature';
 import Hero from './components/Hero';
 import Listing from './components/Listing';
 import Newsroom from './components/Newsroom';
-import PhotoVideo from './components/PhotoVideo';
+import Video from './components/Video';
 import Text from './components/Text';
 import { Abstract } from './components/Abstract';
 import { Accordion } from './components/Accordion/Accordion';
@@ -30,7 +30,7 @@ export {
   Kpi,
   Listing,
   Newsroom,
-  PhotoVideo,
+  Video,
   PreHeader,
   Stats,
   Text,
