+++ src/stories/PhotoVideo/dark.stories.tsx
@@ -1,0 +1,55 @@
+import { type ComponentStory, type ComponentMeta } from '@storybook/react';
+import Video from '../../components/Video';
+import { EIcon } from '../../components/EIcon';
+import { Stack, Typography } from '@mui/material';
+
+export default {
+  title: 'Video/dark',
+  component: Video,
+} as ComponentMeta<typeof Video>;
+
+const Template: ComponentStory<typeof Video> = (args) => <Video {...args} />;
+
+export const Regular = Template.bind({});
+Regular.args = {
+  theme: 'dark',
+  autoplay: true,
+  title: 'Nam eget dui. Etiam rhoncus Maecenas tempus',
+  subtitle:
+    'Lorem ipsum dolor sit amet, consectetuer adipiscing elit. Aenean commodo ligula eget dolor. Aenean massa.',
+  src: 'https://assets.mixkit.co/videos/preview/mixkit-person-typing-on-a-computer-in-detail-4907-large.mp4',
+};
+
+export const Reverse = Template.bind({});
+Reverse.args = {
+  theme: 'dark',
+  reverse: true,
+  autoplay: true,
+  title: 'Nam eget dui. Etiam rhoncus Maecenas tempus',
+  subtitle:
+    'Lorem ipsum dolor sit amet, consectetuer adipiscing elit. Aenean commodo ligula eget dolor. Aenean massa.',
+  src: 'https://assets.mixkit.co/videos/preview/mixkit-person-typing-on-a-computer-in-detail-4907-large.mp4',
+};
+
+export const Full = Template.bind({});
+Full.args = {
+  theme: 'dark',
+  src: 'https://assets.mixkit.co/videos/preview/mixkit-person-typing-on-a-computer-in-detail-4907-large.mp4',
+  autoplay: true,
+  full: true,
+  loop: true,
+};
+
+export const Error = Template.bind({});
+Error.args = {
+  theme: 'dark',
+  full: true,
+  fallback: (
+    <Stack textAlign="center">
+      <EIcon icon="ErrorOutline" color="background.paper" />
+      <Typography variant="h6" color="background.paper">
+        Spiacenti il video non è disponibile in questo momento
+      </Typography>
+    </Stack>
+  ),
+};
